+++ docs/docs-components/siteIndex.js
@@ -110,6 +110,7 @@ const siteIndex: $ReadOnlyArray<siteIndexType> = [
       'Badge',
       'Box',
       'Button',
+      'ButtonLink',
       'ButtonGroup',
       'Callout',
       'Checkbox',
@@ -161,6 +162,7 @@ const siteIndex: $ReadOnlyArray<siteIndexType> = [
       'Sticky',
       'Switch',
       'Table',
+      'TableOfContents',
       'Tabs',
       'Tag',
       'TagData',
@@ -245,7 +247,7 @@ const siteIndex: $ReadOnlyArray<siteIndexType> = [
   },
   {
     sectionName: 'Blog',
-    pages: ['2023 Q1 Newsletter'],
+    pages: ['2023 Q2 Newsletter', '2023 Q1 Newsletter'],
   },
   {
     sectionName: 'Team support',