+++ CHANGELOG.md
@@ -1,3 +1,129 @@
+## 120.3.2 (Aug 9, 2023)
+
+### Patch
+
+- TableOfContents: Add default label ([#3126](https://github.com/pinterest/gestalt/pull/3126)) - [Preview link](https://deploy-preview-3126--gestalt.netlify.app?devexample=true)
+
+## 120.3.1 (Aug 9, 2023)
+
+### Patch
+
+- DateField: fix locales in MM/DD/YYYY format, upgraded package, added unit tests ([#3125](https://github.com/pinterest/gestalt/pull/3125)) - [Preview link](https://deploy-preview-3125--gestalt.netlify.app?devexample=true)
+
+## 120.3.0 (Aug 8, 2023)
+
+### Minor
+
+- TableOfContents: New Component ([#3118](https://github.com/pinterest/gestalt/pull/3118)) - [Preview link](https://deploy-preview-3118--gestalt.netlify.app?devexample=true)
+
+## 120.2.0 (Aug 8, 2023)
+
+### Minor
+
+- ButtonLink: Adding custom codemod to implement component ([#3123](https://github.com/pinterest/gestalt/pull/3123)) - [Preview link](https://deploy-preview-3123--gestalt.netlify.app?devexample=true)
+
+## 120.1.1 (Aug 8, 2023)
+
+### Patch
+
+- Docs: RFC GlobalEventsHandlerProvider and providerAuxData ([#3124](https://github.com/pinterest/gestalt/pull/3124)) - [Preview link](https://deploy-preview-3124--gestalt.netlify.app?devexample=true)
+
+## 120.1.0 (Aug 8, 2023)
+
+### Minor
+
+- GlobalEventsHandlerProvider, DatePicker, DateRange, DateField: add `onMount` handlers ([#3122](https://github.com/pinterest/gestalt/pull/3122)) - [Preview link](https://deploy-preview-3122--gestalt.netlify.app?devexample=true)
+
+## 120.0.2 (Aug 4, 2023)
+
+### Patch
+
+- Docs: newsletter July ([#3120](https://github.com/pinterest/gestalt/pull/3120)) - [Preview link](https://deploy-preview-3120--gestalt.netlify.app?devexample=true)
+
+## 120.0.1 (Aug 3, 2023)
+
+### Patch
+
+- Docs: adding notes on known issues ([#3119](https://github.com/pinterest/gestalt/pull/3119)) - [Preview link](https://deploy-preview-3119--gestalt.netlify.app?devexample=true)
+
+## 120.0.0 (Jul 31, 2023)
+
+### Major
+
+- DatePicker: support controlled/uncontrolled component ([#3115](https://github.com/pinterest/gestalt/pull/3115)) - [Preview link](https://deploy-preview-3115--gestalt.netlify.app?devexample=true)
+
+## 119.5.0 (Jul 31, 2023)
+
+### Minor
+
+- Popover: add "forceDown" to `idealDirection` ([#3111](https://github.com/pinterest/gestalt/pull/3111)) - [Preview link](https://deploy-preview-3111--gestalt.netlify.app?devexample=true)
+
+## 119.4.2 (Jul 31, 2023)
+
+### Patch
+
+- Table: Headercell take full width ([#3116](https://github.com/pinterest/gestalt/pull/3116)) - [Preview link](https://deploy-preview-3116--gestalt.netlify.app?devexample=true)
+
+## 119.4.1 (Jul 28, 2023)
+
+### Patch
+
+- Icon: support rtl to "mute" icon ([#3113](https://github.com/pinterest/gestalt/pull/3113)) - [Preview link](https://deploy-preview-3113--gestalt.netlify.app?devexample=true)
+
+## 119.4.0 (Jul 28, 2023)
+
+### Minor
+
+- IconButton: support to `data-test-id` ([#3101](https://github.com/pinterest/gestalt/pull/3101)) - [Preview link](https://deploy-preview-3101--gestalt.netlify.app?devexample=true)
+
+## 119.3.0 (Jul 27, 2023)
+
+### Minor
+
+- DateField: Add locale format translations ([#3103](https://github.com/pinterest/gestalt/pull/3103)) - [Preview link](https://deploy-preview-3103--gestalt.netlify.app?devexample=true)
+
+## 119.2.5 (Jul 27, 2023)
+
+### Patch
+
+- Docs: Switch replacing images on the iOS doc ([#3109](https://github.com/pinterest/gestalt/pull/3109)) - [Preview link](https://deploy-preview-3109--gestalt.netlify.app?devexample=true)
+
+## 119.2.4 (Jul 27, 2023)
+
+### Patch
+
+- Docs: Switch iOS mobile doc clean up ([#3104](https://github.com/pinterest/gestalt/pull/3104)) - [Preview link](https://deploy-preview-3104--gestalt.netlify.app?devexample=true)
+
+## 119.2.3 (Jul 27, 2023)
+
+### Patch
+
+- Docs: New and updated component svgs ([#3108](https://github.com/pinterest/gestalt/pull/3108)) - [Preview link](https://deploy-preview-3108--gestalt.netlify.app?devexample=true)
+
+## 119.2.2 (Jul 27, 2023)
+
+### Patch
+
+- Docs: development mode switch shown when activated from url ([#3107](https://github.com/pinterest/gestalt/pull/3107)) - [Preview link](https://deploy-preview-3107--gestalt.netlify.app?devexample=true)
+
+## 119.2.1 (Jul 26, 2023)
+
+### Patch
+
+- Docs: Added iOS components to the H2 roadmap ([#3102](https://github.com/pinterest/gestalt/pull/3102)) - [Preview link](https://deploy-preview-3102--gestalt.netlify.app?devexample=true)
+
+## 119.2.0 (Jul 26, 2023)
+
+### Minor
+
+- ButtonLink: new component forked from Button ([#3093](https://github.com/pinterest/gestalt/pull/3093)) - [Preview link](https://deploy-preview-3093--gestalt.netlify.app?devexample=true)
+
+## 119.1.5 (Jul 25, 2023)
+
+### Patch
+
+- Docs: Q2 newsletter addition ([#3061](https://github.com/pinterest/gestalt/pull/3061)) - [Preview link](https://deploy-preview-3061--gestalt.netlify.app?devexample=true)
+
 ## 119.1.4 (Jul 24, 2023)
 
 ### Patch